requirements.txt
fastapi
pytest
cookiecutter-data-science
mkdocs
python-dotenv
pydantic
uvicorn